auth_server/requirements.txt
Django
django-environ
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-decouple
dj-database-url
grpcio
grpcio-tools